docs/docs_generator/requirements.txt
# Documentation Generation and Conversion
click==8.3.0
lxml==6.0.2
pypandoc==1.15
python-docx==1.2.0
typing_extensions==4.15.0
progressbar2==4.5.0
